+++ src/__init__.py
@@ -1,1 +1,9 @@
+# src/__init__.py
+
 from .config_utils import load_config
+from .data_utils import EAPDataset, prepare_bias_corrupt
+from .utils import get_metric
+from .circuit_baseline import evaluate_baseline
+from .circuit_eap import attribute
+from .circuit_eval import evaluate_graph
+from .graph import Graph, load_graph_from_json
